Move servlet-common internal package under servlet.common.internal
Aligns the internal package layout of the servlet-common library module with the module-name <-> package-name convention by relocating io.opentelemetry.instrumentation.servlet.internal to io.opentelemetry.instrumentation.servlet.common.internal.

These classes are marked internal (not part of the public API) and are consumed only by other instrumentation modules in this repository, so updating the import sites is sufficient.

Also drops the wholesale servlet-common module skip in the package-name checker, since the package layout now satisfies the prefix check via the existing no-base-version allow-list.
