+++ .github/scripts/check-package-names.sh
@@ -13,9 +13,6 @@ for dir in $(find instrumentation -name "*.java" | grep library/src/main/java | 
   if [[ "$module_name" == "runtime-telemetry" ]]; then
     continue
   fi
-  if [[ "$module_name" == "servlet-common" ]]; then
-    continue
-  fi
   if [[ "$module_name" == "graphql-java-common-12.0" ]]; then
     continue
   fi
@@ -40,8 +37,9 @@ for dir in $(find instrumentation -name "*.java" | grep library/src/main/java | 
   # - netty-common
   # - oshi
   # - resources
+  # - servlet-common
   # - servlet-common-javax
-  if [[ ! "$module_name" =~ [0-9]$ && "$module_name" != "jdbc" && "$module_name" != "lettuce-common" && "$module_name" != "netty-common" && "$module_name" != "oshi" && "$module_name" != "resources" && "$module_name" != "servlet-common-javax" ]]; then
+  if [[ ! "$module_name" =~ [0-9]$ && "$module_name" != "jdbc" && "$module_name" != "lettuce-common" && "$module_name" != "netty-common" && "$module_name" != "oshi" && "$module_name" != "resources" && "$module_name" != "servlet-common" && "$module_name" != "servlet-common-javax" ]]; then
     echo "module name doesn't have a base version: $dir"
     exit 1
   fi
