+++ instrumentation/jetty/jetty-common-8.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jetty/common/v8_0/JettyHelper.java
@@ -8,9 +8,9 @@ package io.opentelemetry.javaagent.instrumentation.jetty.common.v8_0;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
-import io.opentelemetry.instrumentation.servlet.internal.ServletAccessor;
-import io.opentelemetry.instrumentation.servlet.internal.ServletRequestContext;
-import io.opentelemetry.instrumentation.servlet.internal.ServletResponseContext;
+import io.opentelemetry.instrumentation.servlet.common.internal.ServletAccessor;
+import io.opentelemetry.instrumentation.servlet.common.internal.ServletRequestContext;
+import io.opentelemetry.instrumentation.servlet.common.internal.ServletResponseContext;
 import io.opentelemetry.javaagent.instrumentation.servlet.common.ServletHelper;
 import javax.annotation.Nullable;
