+++ instrumentation/servlet/servlet-common/library/src/main/java/io/opentelemetry/instrumentation/servlet/common/internal/ServletErrorCauseExtractor.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.servlet.internal;
+package io.opentelemetry.instrumentation.servlet.common.internal;
 
 import io.opentelemetry.instrumentation.api.instrumenter.ErrorCauseExtractor;
